services/inventory-service/pyproject.toml
[project]
name = "inventory-service"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.13",
    "uvicorn==0.34.0",
    "httpx==0.28.1",
    "openfeature-sdk==0.8.4",
    "openfeature-provider-flagd==0.3.0",
]
